requirements.txt
Django==5.1.6
djangorestframework-simplejwt
PyJWT
gunicorn
